api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
supabase==2.0.2
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
